add CORELoan and CORELoanFactory contract

+++ contracts/SimpleStorage.sol
@@ -1,14 +1,15 @@
-// SPDX-License-Identifier: GPL-3.0
+// SPDX-License-Identifier: MIT
+
 pragma solidity >=0.4.16 <0.8.0;
 
 contract SimpleStorage {
-    uint storedData;
+    uint256 storedData;
 
-    function set(uint x) public {
+    function set(uint256 x) public {
         storedData = x;
     }
 
-    function get() public view returns (uint) {
+    function get() public view returns (uint256) {
         return storedData;
     }
 }
